feat(plan-200): support up --wait on the vz backend (verdict-capture)
Extends the one-shot `up --wait` verdict-capture from libkrun to vz, the last
WS-B live-validation gap. The libkrun egress matrix is already verified 0/3/2
(#1083); vz uses a connected socketpair so the #1083 recvfrom-source fix does
not apply to it.

- New shared mvm_backend::workload_wait module (read_exit_status_from +
  wait_for_workload_exit): state-dir based, backend-agnostic. The supervisor
  persists <vm_state_dir>/workload.exit before poweroff (ack-ordered), so the
  file's presence is the reliable "workload finished" signal; no PID-liveness
  poll (a reused PID on a busy host spins forever); bounded -> UNKNOWN.
- libkrun's VmBackend::wait now delegates to the shared helper (was a private
  copy of the same loop).
- vz VmBackend::wait delegates to the same helper (the vz supervisor already
  persists workload.exit, vz_objc.rs).
- Relax both up.rs --wait gates from == "libkrun" to matches!(.., "libkrun"|"vz").

Additive + safe: if a vz workload fails to boot, `up` errors before reaching
`wait`, exactly as today. Unit tests cover the shared helper (nonzero/zero/
absent/staged). Live-vz 0/3/2 is not yet proven end-to-end: it surfaced a
separate vz one-shot workload-boot failure (supervisor exits before writing the
PID file, empty console.log) that fails at boot, before `wait` — filed as a
deferred follow-up. The wait() logic rides the exact shared path proven live
0/3/2 on libkrun.
