+++ specs/REFACTOR-STATUS.md
@@ -22,6 +22,8 @@
 
 **Additional 2026-06-19 — Plan 200 WS-B follow-up (d), live matrix exposed + fixed a real libkrun bridge bug:** The deferred live matrix returned verdict 3 for ALL policies (incl. unrestricted). The chain-signed audit proved the bridge policy was correct but the internet→guest return frames never reached the guest: `run_libkrun_gvproxy_bridge` replied to the recvfrom source instead of libkrun's derived `<listen>-krun.sock` reply listener (not a usable reply target on macOS), silently dropping every inbound frame. Fixed (reply to the derived path; matches the rvproxy native-gateway precedent); the duplex test was rewritten to model real libkrun so it catches the regression. **Live matrix now passes 0 / 3 / 2** on current origin/main. follow-up (c)'s bound-client test had masked it — only the live VM run caught it. Branch `fix/plan-200-libkrun-bridge-ingress-reply`.
 
+**Additional 2026-06-19 — Plan 200 WS-B follow-up (e), Vz `up --wait` verdict-capture implemented:** Extracted the `wait()` poll into a shared backend-agnostic `mvm_backend::workload_wait` module; libkrun + Vz backends both delegate to it (the vz supervisor already persists `workload.exit`); relaxed both `up.rs --wait` gates to `libkrun|vz`. Unit-tested + clippy/fmt/spec/linux-cross-compile clean. **Live-vz `0/3/2` not yet proven** — surfaced a separate vz one-shot workload-boot failure (`supervisor exited before writing PID file`, empty console.log, fails at boot before `wait`), filed as a deferred follow-up. The `wait()` logic rides the shared path already proven live `0/3/2` on libkrun. Branch `feat/plan-200-vz-up-wait`.
+
 **Additional 2026-06-19 — Plan 195 closeout:** Builder-VM fingerprint narrowing is complete. `builder_vm_source_fingerprint` ignores unrelated workspace `Cargo.lock` churn, keeps flake / `nix/lib` / embedded-host-binary byte identity as invalidation inputs, and `mvm-cli/build.rs` watches `Cargo.lock` + `crates/mvm-build/src` so those embedded bytes rebuild when their real inputs change. Closeout fixed stale comments that still described the removed lockfile layer and recorded verification: focused fingerprint tests (6/6), real non-stub embedded-binary build, `mvm-build/src` touch causing `cargo zigbuild` rerun, full `mvm-cli` nextest (1088 passed, 1 skipped), clippy, and fmt. The live `mvmctl dev up` manual was not re-run in this session because runtime commands must run inside the project builder VM and this checkout has no wrapper for that; deterministic cache-key tests cover the hit/mismatch behavior without booting a VM.
 
 **Additional 2026-06-19 — Plan 200 auth-proof live follow-up:** the Firecracker-host proof now reaches guest boot on branch-local `mvmctl`: a dev-profile `alpine:latest` persistent machine with `[auth].ssh_agent = true` starts, signed receipt/dry-run/audit surfaces report `ssh-agent-socket`, and direct raw `SSH_AGENTC_REQUEST_IDENTITIES` probes to the throwaway host agent and spawned per-machine proxy UDS both return an SSH-agent identities answer. The proof remains unclaimed because the in-guest raw probe copied to `/tmp/mvm-agent-probe-c` reaches `/run/mvm/ssh-agent.sock` but reads `Connection reset by peer`, narrowing the blocker to Firecracker guest-to-host host-listen forwarding for port 5301. Follow-up code now routes Firecracker SSH-agent proxy traffic through the per-port runtime UDS (`vm_vsock_port_socket(..., 5301)`) instead of raw host AF_VSOCK, with unit coverage pinning backend socket transport; the live in-guest raw probe still needs rerun before the smoke is claimed. Runtime packet enforcement now drops inbound SSH identification banners on any TCP port via `ssh-banner-protocol-deny` and reverse-flow kill matching, but a live non-22 guest witness is still needed before claiming the SSH-banner smoke.
